+++ sdk/go/aws/apigateway/integrationResponse.go
@@ -144,7 +144,7 @@ type integrationResponseArgs struct {
 	// A map specifying the templates used to transform the integration response body
 	ResponseTemplates map[string]string `pulumi:"responseTemplates"`
 	// The ID of the associated REST API
-	RestApi string `pulumi:"restApi"`
+	RestApi interface{} `pulumi:"restApi"`
 	// Specifies the regular expression pattern used to choose
 	// an integration response based on the response from the backend. Setting this to `-` makes the integration the default one.
 	// If the backend is an `AWS` Lambda function, the AWS Lambda function error header is matched.
@@ -168,7 +168,7 @@ type IntegrationResponseArgs struct {
 	// A map specifying the templates used to transform the integration response body
 	ResponseTemplates pulumi.StringMapInput
 	// The ID of the associated REST API
-	RestApi pulumi.StringInput
+	RestApi pulumi.Input
 	// Specifies the regular expression pattern used to choose
 	// an integration response based on the response from the backend. Setting this to `-` makes the integration the default one.
 	// If the backend is an `AWS` Lambda function, the AWS Lambda function error header is matched.